Move (ResultPrinter), refs 3512
