+++ src/Aliases.php
@@ -4,6 +4,12 @@
  * SemanticMediaWiki compatibility aliases for classes that got moved into the SMW namespace
  */
 
+// 3.1
+class_alias( \SMW\Query\ResultPrinters\RdfResultPrinter::class, 'SMWRDFResultPrinter' );
+class_alias( \SMW\Query\ResultPrinters\EmbeddedResultPrinter::class, 'SMWEmbeddedResultPrinter' );
+class_alias( \SMW\Query\ResultPrinters\DsvResultPrinter::class, 'SMWDSVResultPrinter' );
+class_alias( \SMW\Query\ResultPrinters\AggregatablePrinter::class, 'SMWAggregatablePrinter' );
+
 // 3.0
 class_alias( \SMW\MediaWiki\Deferred\CallableUpdate::class, 'SMW\DeferredCallableUpdate' );
 class_alias( \SMW\Parser\InTextAnnotationParser::class, 'SMW\InTextAnnotationParser' );
@@ -40,11 +46,7 @@ class_alias( \SMW\Query\ResultPrinters\TableResultPrinter::class, 'SMWTableResul
 
 // 2.0
 class_alias( \SMW\Query\ResultPrinters\FileExportPrinter::class, 'SMWExportPrinter' );
-class_alias( \SMW\AggregatablePrinter::class, 'SMWAggregatablePrinter' );
 class_alias( \SMW\Query\ResultPrinters\CategoryResultPrinter::class, 'SMWCategoryResultPrinter' );
-class_alias( \SMW\DsvResultPrinter::class, 'SMWDSVResultPrinter' );
-class_alias( \SMW\EmbeddedResultPrinter::class, 'SMWEmbeddedResultPrinter' );
-class_alias( \SMW\RdfResultPrinter::class, 'SMWRDFResultPrinter' );
 class_alias( \SMW\ListResultPrinter::class, 'SMWListResultPrinter' );
 
 // 2.0
