+++ src/Query/ResultPrinters/AggregatablePrinter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SMW;
+namespace SMW\Query\ResultPrinters;
 
 use SMWDataItem as DataItem;
 use SMWQueryResult as QueryResult;
